# REST Server/target/docs/apidocs/ui/swagger.json
{
  "swagger": "2.0",
  "info" : {
    "description" : "<h1>MavenServer<\/h1>",
    "version" : "1.3",
    "title" : "MavenServer"
  },
  "schemes" : [],
...
  "definitions" : {
...
    "json_JsonObject" : {
      "type" : "object",
      "title" : "JsonObject",
      "allOf" : [
        {
          "$ref" : "#/definitions/json_JsonElement"
        },
        {
        }
      ],
      "example" : {
  "asString" : "...",
  "asDouble" : 12345.0,
  "asFloat" : 12345.0,
  "asJsonNull" : {
    "asString" : "...",
    "asDouble" : 12345.0,
    "asFloat" : 12345.0,
    "asJsonNull" : { },
    "jsonPrimitive" : true,
    "asBoolean" : true,
    "asNumber" : { },
    "asJsonArray" : {
      "asString" : "...",
      "asDouble" : 12345.0,
      "asFloat" : 12345.0,
      "asJsonNull" : { },
      "jsonPrimitive" : true,
      "asBoolean" : true,
      "asNumber" : { },
      "asJsonArray" : { },
      "asLong" : 12345,
      "jsonNull" : true,
      "asJsonPrimitive" : { },
      "asInt" : 12345,
      "asByte" : 12345,
      "asBigDecimal" : 12345.0,
      "asShort" : 12345,
      "asBigInteger" : 12345,
      "jsonArray" : true,
      "asJsonObject" : { },
      "asCharacter" : { },
      "jsonObject" : true
    },
    "asLong" : 12345,
    "jsonNull" : true,
    "asJsonPrimitive" : {
      "string" : true,
      "number" : true,
      "boolean" : true,
      "asString" : "...",
      "asDouble" : 12345.0,
      "asFloat" : 12345.0,
      "asJsonNull" : { },
      "jsonPrimitive" : true,
      "asBoolean" : true,
      "asNumber" : { },
      "asJsonArray" : { },
      "asLong" : 12345,
      "jsonNull" : true,
      "asJsonPrimitive" : { },
      "asInt" : 12345,
      "asByte" : 12345,
      "asBigDecimal" : 12345.0,
      "asShort" : 12345,
      "asBigInteger" : 12345,
      "jsonArray" : true,
      "asJsonObject" : { },
      "asCharacter" : { },
      "jsonObject" : true
    },
    "asInt" : 12345,
    "asByte" : 12345,
    "asBigDecimal" : 12345.0,
    "asShort" : 12345,
    "asBigInteger" : 12345,
    "jsonArray" : true,
    "asJsonObject" : {
      "asString" : "...",
      "asDouble" : 12345.0,
      "asFloat" : 12345.0,
      "asJsonNull" : { },
      "jsonPrimitive" : true,
      "asBoolean" : true,
      "asNumber" : { },
      "asJsonArray" : { },
      "asLong" : 12345,
      "jsonNull" : true,
      "asJsonPrimitive" : { },
      "asInt" : 12345,
      "asByte" : 12345,
      "asBigDecimal" : 12345.0,
      "asShort" : 12345,
      "asBigInteger" : 12345,
      "jsonArray" : true,
      "asJsonObject" : { },
      "asCharacter" : { },
      "jsonObject" : true
    },
    "asCharacter" : { },
    "jsonObject" : true
  },
  "jsonPrimitive" : true,
  "asBoolean" : true,
  "asNumber" : { },
  "asJsonArray" : {
    "asString" : "...",
    "asDouble" : 12345.0,
    "asFloat" : 12345.0,
    "asJsonNull" : {
      "asString" : "...",
      "asDouble" : 12345.0,
      "asFloat" : 12345.0,
      "asJsonNull" : { },
      "jsonPrimitive" : true,
      "asBoolean" : true,
      "asNumber" : { },
      "asJsonArray" : { },
      "asLong" : 12345,
      "jsonNull" : true,
      "asJsonPrimitive" : { },
      "asInt" : 12345,
      "asByte" : 12345,
      "asBigDecimal" : 12345.0,
      "asShort" : 12345,
      "asBigInteger" : 12345,
      "jsonArray" : true,
      "asJsonObject" : { },
      "asCharacter" : { },
      "jsonObject" : true
    },
    "jsonPrimitive" : true,
    "asBoolean" : true,
    "asNumber" : { },
    "asJsonArray" : { },
    "asLong" : 12345,
    "jsonNull" : true,
    "asJsonPrimitive" : {
      "string" : true,
      "number" : true,
      "boolean" : true,
      "asString" : "...",
      "asDouble" : 12345.0,
      "asFloat" : 12345.0,
      "asJsonNull" : { },
      "jsonPrimitive" : true,
      "asBoolean" : true,
      "asNumber" : { },
      "asJsonArray" : { },
      "asLong" : 12345,
      "jsonNull" : true,
      "asJsonPrimitive" : { },
      "asInt" : 12345,
      "asByte" : 12345,
      "asBigDecimal" : 12345.0,
      "asShort" : 12345,
      "asBigInteger" : 12345,
      "jsonArray" : true,
      "asJsonObject" : { },
      "asCharacter" : { },
      "jsonObject" : true
    },
    "asInt" : 12345,
    "asByte" : 12345,
    "asBigDecimal" : 12345.0,
    "asShort" : 12345,
    "asBigInteger" : 12345,
    "jsonArray" : true,
    "asJsonObject" : {
      "asString" : "...",
      "asDouble" : 12345.0,
      "asFloat" : 12345.0,
      "asJsonNull" : { },
      "jsonPrimitive" : true,
      "asBoolean" : true,
      "asNumber" : { },
      "asJsonArray" : { },
      "asLong" : 12345,
      "jsonNull" : true,
      "asJsonPrimitive" : { },
      "asInt" : 12345,
      "asByte" : 12345,
      "asBigDecimal" : 12345.0,
      "asShort" : 12345,
      "asBigInteger" : 12345,
      "jsonArray" : true,
      "asJsonObject" : { },
      "asCharacter" : { },
      "jsonObject" : true
    },
    "asCharacter" : { },
    "jsonObject" : true
  },
  "asLong" : 12345,
  "jsonNull" : true,
  "asJsonPrimitive" : {
    "string" : true,
    "number" : true,
    "boolean" : true,
    "asString" : "...",
    "asDouble" : 12345.0,
    "asFloat" : 12345.0,
    "asJsonNull" : {
      "asString" : "...",
      "asDouble" : 12345.0,
      "asFloat" : 12345.0,
      "asJsonNull" : { },
      "jsonPrimitive" : true,
      "asBoolean" : true,
      "asNumber" : { },
      "asJsonArray" : { },
      "asLong" : 12345,
      "jsonNull" : true,
      "asJsonPrimitive" : { },
      "asInt" : 12345,
      "asByte" : 12345,
      "asBigDecimal" : 12345.0,
      "asShort" : 12345,
      "asBigInteger" : 12345,
      "jsonArray" : true,
      "asJsonObject" : { },
      "asCharacter" : { },
      "jsonObject" : true
    },
    "jsonPrimitive" : true,
    "asBoolean" : true,
    "asNumber" : { },
    "asJsonArray" : {
      "asString" : "...",
      "asDouble" : 12345.0,
      "asFloat" : 12345.0,
      "asJsonNull" : { },
      "jsonPrimitive" : true,
      "asBoolean" : true,
      "asNumber" : { },
      "asJsonArray" : { },
      "asLong" : 12345,
      "jsonNull" : true,
      "asJsonPrimitive" : { },
      "asInt" : 12345,
      "asByte" : 12345,
      "asBigDecimal" : 12345.0,
      "asShort" : 12345,
      "asBigInteger" : 12345,
      "jsonArray" : true,
      "asJsonObject" : { },
      "asCharacter" : { },
      "jsonObject" : true
    },
    "asLong" : 12345,
    "jsonNull" : true,
    "asJsonPrimitive" : { },
    "asInt" : 12345,
    "asByte" : 12345,
    "asBigDecimal" : 12345.0,
    "asShort" : 12345,
    "asBigInteger" : 12345,
    "jsonArray" : true,
    "asJsonObject" : {
      "asString" : "...",
      "asDouble" : 12345.0,
      "asFloat" : 12345.0,
      "asJsonNull" : { },
      "jsonPrimitive" : true,
      "asBoolean" : true,
      "asNumber" : { },
      "asJsonArray" : { },
      "asLong" : 12345,
      "jsonNull" : true,
      "asJsonPrimitive" : { },
      "asInt" : 12345,
      "asByte" : 12345,
      "asBigDecimal" : 12345.0,
      "asShort" : 12345,
      "asBigInteger" : 12345,
      "jsonArray" : true,
      "asJsonObject" : { },
      "asCharacter" : { },
      "jsonObject" : true
    },
    "asCharacter" : { },
    "jsonObject" : true
  },
  "asInt" : 12345,
  "asByte" : 12345,
  "asBigDecimal" : 12345.0,
  "asShort" : 12345,
  "asBigInteger" : 12345,
  "jsonArray" : true,
  "asJsonObject" : {
    "asString" : "...",
    "asDouble" : 12345.0,
    "asFloat" : 12345.0,
    "asJsonNull" : {
      "asString" : "...",
      "asDouble" : 12345.0,
      "asFloat" : 12345.0,
      "asJsonNull" : { },
      "jsonPrimitive" : true,
      "asBoolean" : true,
      "asNumber" : { },
      "asJsonArray" : { },
      "asLong" : 12345,
      "jsonNull" : true,
      "asJsonPrimitive" : { },
      "asInt" : 12345,
      "asByte" : 12345,
      "asBigDecimal" : 12345.0,
      "asShort" : 12345,
      "asBigInteger" : 12345,
      "jsonArray" : true,
      "asJsonObject" : { },
      "asCharacter" : { },
      "jsonObject" : true
    },
    "jsonPrimitive" : true,
    "asBoolean" : true,
    "asNumber" : { },
    "asJsonArray" : {
      "asString" : "...",
      "asDouble" : 12345.0,
      "asFloat" : 12345.0,
      "asJsonNull" : { },
      "jsonPrimitive" : true,
      "asBoolean" : true,
      "asNumber" : { },
      "asJsonArray" : { },
      "asLong" : 12345,
      "jsonNull" : true,
      "asJsonPrimitive" : { },
      "asInt" : 12345,
      "asByte" : 12345,
      "asBigDecimal" : 12345.0,
      "asShort" : 12345,
      "asBigInteger" : 12345,
      "jsonArray" : true,
      "asJsonObject" : { },
      "asCharacter" : { },
      "jsonObject" : true
    },
    "asLong" : 12345,
    "jsonNull" : true,
    "asJsonPrimitive" : {
      "string" : true,
      "number" : true,
      "boolean" : true,
      "asString" : "...",
      "asDouble" : 12345.0,
      "asFloat" : 12345.0,
      "asJsonNull" : { },
      "jsonPrimitive" : true,
      "asBoolean" : true,
      "asNumber" : { },
      "asJsonArray" : { },
      "asLong" : 12345,
      "jsonNull" : true,
      "asJsonPrimitive" : { },
      "asInt" : 12345,
      "asByte" : 12345,
      "asBigDecimal" : 12345.0,
      "asShort" : 12345,
      "asBigInteger" : 12345,
      "jsonArray" : true,
      "asJsonObject" : { },
      "asCharacter" : { },
      "jsonObject" : true
    },
    "asInt" : 12345,
    "asByte" : 12345,
    "asBigDecimal" : 12345.0,
    "asShort" : 12345,
    "asBigInteger" : 12345,
    "jsonArray" : true,
    "asJsonObject" : { },
    "asCharacter" : { },
    "jsonObject" : true
  },
  "asCharacter" : { },
  "jsonObject" : true
},
      "description" : ""
    }
...
  },
  "paths": {
...
    "\/stores/{store}/items/add/{id}" : {
      "put" : {
        "tags" : [ "Stores" ],
        "description" : "",
        "operationId" : "resource_Stores_addItemToCart_PUT",
        "consumes" : [ "text/plain" ],
        "produces" : [ "application/json" ],
        "parameters" : [
          {
            "name" : "id",
            "in" : "path",
            "required" : true,
            "type" : "number",
            "description" : ""
          },
          {
            "name" : "store",
            "in" : "path",
            "required" : true,
            "type" : "integer",
            "description" : ""
          },
          {
            "name" : "body",
            "in" : "body",
            "schema" : {
"description" : "",
"type" : "string"
            },
            "description" : ""
          }
        ],
        "responses" : {
          "204" : {
            "schema" : {
"description" : "",
"$ref" : "#/definitions/json_JsonObject"
            },
            "headers" : {
            },
            "examples" : {
              "application/json" : {
  "asString" : "...",
  "asDouble" : 12345.0,
  "asFloat" : 12345.0,
  "asJsonNull" : {
    "asString" : "...",
    "asDouble" : 12345.0,
    "asFloat" : 12345.0,
    "asJsonNull" : { },
    "jsonPrimitive" : true,
    "asBoolean" : true,
    "asNumber" : { },
    "asJsonArray" : {
      "asString" : "...",
      "asDouble" : 12345.0,
      "asFloat" : 12345.0,
      "asJsonNull" : { },
      "jsonPrimitive" : true,
      "asBoolean" : true,
      "asNumber" : { },
      "asJsonArray" : { },
      "asLong" : 12345,
      "jsonNull" : true,
      "asJsonPrimitive" : { },
      "asInt" : 12345,
      "asByte" : 12345,
      "asBigDecimal" : 12345.0,
      "asShort" : 12345,
      "asBigInteger" : 12345,
      "jsonArray" : true,
      "asJsonObject" : { },
      "asCharacter" : { },
      "jsonObject" : true
    },
    "asLong" : 12345,
    "jsonNull" : true,
    "asJsonPrimitive" : {
      "string" : true,
      "number" : true,
      "boolean" : true,
      "asString" : "...",
      "asDouble" : 12345.0,
      "asFloat" : 12345.0,
      "asJsonNull" : { },
      "jsonPrimitive" : true,
      "asBoolean" : true,
      "asNumber" : { },
      "asJsonArray" : { },
      "asLong" : 12345,
      "jsonNull" : true,
      "asJsonPrimitive" : { },
      "asInt" : 12345,
      "asByte" : 12345,
      "asBigDecimal" : 12345.0,
      "asShort" : 12345,
      "asBigInteger" : 12345,
      "jsonArray" : true,
      "asJsonObject" : { },
      "asCharacter" : { },
      "jsonObject" : true
    },
    "asInt" : 12345,
    "asByte" : 12345,
    "asBigDecimal" : 12345.0,
    "asShort" : 12345,
    "asBigInteger" : 12345,
    "jsonArray" : true,
    "asJsonObject" : {
      "asString" : "...",
      "asDouble" : 12345.0,
      "asFloat" : 12345.0,
      "asJsonNull" : { },
      "jsonPrimitive" : true,
      "asBoolean" : true,
      "asNumber" : { },
      "asJsonArray" : { },
      "asLong" : 12345,
      "jsonNull" : true,
      "asJsonPrimitive" : { },
      "asInt" : 12345,
      "asByte" : 12345,
      "asBigDecimal" : 12345.0,
      "asShort" : 12345,
      "asBigInteger" : 12345,
      "jsonArray" : true,
      "asJsonObject" : { },
      "asCharacter" : { },
      "jsonObject" : true
    },
    "asCharacter" : { },
    "jsonObject" : true
  },
  "jsonPrimitive" : true,
  "asBoolean" : true,
  "asNumber" : { },
  "asJsonArray" : {
    "asString" : "...",
    "asDouble" : 12345.0,
    "asFloat" : 12345.0,
    "asJsonNull" : {
      "asString" : "...",
      "asDouble" : 12345.0,
      "asFloat" : 12345.0,
      "asJsonNull" : { },
      "jsonPrimitive" : true,
      "asBoolean" : true,
      "asNumber" : { },
      "asJsonArray" : { },
      "asLong" : 12345,
      "jsonNull" : true,
      "asJsonPrimitive" : { },
      "asInt" : 12345,
      "asByte" : 12345,
      "asBigDecimal" : 12345.0,
      "asShort" : 12345,
      "asBigInteger" : 12345,
      "jsonArray" : true,
      "asJsonObject" : { },
      "asCharacter" : { },
      "jsonObject" : true
    },
    "jsonPrimitive" : true,
    "asBoolean" : true,
    "asNumber" : { },
    "asJsonArray" : { },
    "asLong" : 12345,
    "jsonNull" : true,
    "asJsonPrimitive" : {
      "string" : true,
      "number" : true,
      "boolean" : true,
      "asString" : "...",
      "asDouble" : 12345.0,
      "asFloat" : 12345.0,
      "asJsonNull" : { },
      "jsonPrimitive" : true,
      "asBoolean" : true,
      "asNumber" : { },
      "asJsonArray" : { },
      "asLong" : 12345,
      "jsonNull" : true,
      "asJsonPrimitive" : { },
      "asInt" : 12345,
      "asByte" : 12345,
      "asBigDecimal" : 12345.0,
      "asShort" : 12345,
      "asBigInteger" : 12345,
      "jsonArray" : true,
      "asJsonObject" : { },
      "asCharacter" : { },
      "jsonObject" : true
    },
    "asInt" : 12345,
    "asByte" : 12345,
    "asBigDecimal" : 12345.0,
    "asShort" : 12345,
    "asBigInteger" : 12345,
    "jsonArray" : true,
    "asJsonObject" : {
      "asString" : "...",
      "asDouble" : 12345.0,
      "asFloat" : 12345.0,
      "asJsonNull" : { },
      "jsonPrimitive" : true,
      "asBoolean" : true,
      "asNumber" : { },
      "asJsonArray" : { },
      "asLong" : 12345,
      "jsonNull" : true,
      "asJsonPrimitive" : { },
      "asInt" : 12345,
      "asByte" : 12345,
      "asBigDecimal" : 12345.0,
      "asShort" : 12345,
      "asBigInteger" : 12345,
      "jsonArray" : true,
      "asJsonObject" : { },
      "asCharacter" : { },
      "jsonObject" : true
    },
    "asCharacter" : { },
    "jsonObject" : true
  },
  "asLong" : 12345,
  "jsonNull" : true,
  "asJsonPrimitive" : {
    "string" : true,
    "number" : true,
    "boolean" : true,
    "asString" : "...",
    "asDouble" : 12345.0,
    "asFloat" : 12345.0,
    "asJsonNull" : {
      "asString" : "...",
      "asDouble" : 12345.0,
      "asFloat" : 12345.0,
      "asJsonNull" : { },
      "jsonPrimitive" : true,
      "asBoolean" : true,
      "asNumber" : { },
      "asJsonArray" : { },
      "asLong" : 12345,
      "jsonNull" : true,
      "asJsonPrimitive" : { },
      "asInt" : 12345,
      "asByte" : 12345,
      "asBigDecimal" : 12345.0,
      "asShort" : 12345,
      "asBigInteger" : 12345,
      "jsonArray" : true,
      "asJsonObject" : { },
      "asCharacter" : { },
      "jsonObject" : true
    },
    "jsonPrimitive" : true,
    "asBoolean" : true,
    "asNumber" : { },
    "asJsonArray" : {
      "asString" : "...",
      "asDouble" : 12345.0,
      "asFloat" : 12345.0,
      "asJsonNull" : { },
      "jsonPrimitive" : true,
      "asBoolean" : true,
      "asNumber" : { },
      "asJsonArray" : { },
      "asLong" : 12345,
      "jsonNull" : true,
      "asJsonPrimitive" : { },
      "asInt" : 12345,
      "asByte" : 12345,
      "asBigDecimal" : 12345.0,
      "asShort" : 12345,
      "asBigInteger" : 12345,
      "jsonArray" : true,
      "asJsonObject" : { },
      "asCharacter" : { },
      "jsonObject" : true
    },
    "asLong" : 12345,
    "jsonNull" : true,
    "asJsonPrimitive" : { },
    "asInt" : 12345,
    "asByte" : 12345,
    "asBigDecimal" : 12345.0,
    "asShort" : 12345,
    "asBigInteger" : 12345,
    "jsonArray" : true,
    "asJsonObject" : {
      "asString" : "...",
      "asDouble" : 12345.0,
      "asFloat" : 12345.0,
      "asJsonNull" : { },
      "jsonPrimitive" : true,
      "asBoolean" : true,
      "asNumber" : { },
      "asJsonArray" : { },
      "asLong" : 12345,
      "jsonNull" : true,
      "asJsonPrimitive" : { },
      "asInt" : 12345,
      "asByte" : 12345,
      "asBigDecimal" : 12345.0,
      "asShort" : 12345,
      "asBigInteger" : 12345,
      "jsonArray" : true,
      "asJsonObject" : { },
      "asCharacter" : { },
      "jsonObject" : true
    },
    "asCharacter" : { },
    "jsonObject" : true
  },
  "asInt" : 12345,
  "asByte" : 12345,
  "asBigDecimal" : 12345.0,
  "asShort" : 12345,
  "asBigInteger" : 12345,
  "jsonArray" : true,
  "asJsonObject" : {
    "asString" : "...",
    "asDouble" : 12345.0,
    "asFloat" : 12345.0,
    "asJsonNull" : {
      "asString" : "...",
      "asDouble" : 12345.0,
      "asFloat" : 12345.0,
      "asJsonNull" : { },
      "jsonPrimitive" : true,
      "asBoolean" : true,
      "asNumber" : { },
      "asJsonArray" : { },
      "asLong" : 12345,
      "jsonNull" : true,
      "asJsonPrimitive" : { },
      "asInt" : 12345,
      "asByte" : 12345,
      "asBigDecimal" : 12345.0,
      "asShort" : 12345,
      "asBigInteger" : 12345,
      "jsonArray" : true,
      "asJsonObject" : { },
      "asCharacter" : { },
      "jsonObject" : true
    },
    "jsonPrimitive" : true,
    "asBoolean" : true,
    "asNumber" : { },
    "asJsonArray" : {
      "asString" : "...",
      "asDouble" : 12345.0,
      "asFloat" : 12345.0,
      "asJsonNull" : { },
      "jsonPrimitive" : true,
      "asBoolean" : true,
      "asNumber" : { },
      "asJsonArray" : { },
      "asLong" : 12345,
      "jsonNull" : true,
      "asJsonPrimitive" : { },
      "asInt" : 12345,
      "asByte" : 12345,
      "asBigDecimal" : 12345.0,
      "asShort" : 12345,
      "asBigInteger" : 12345,
      "jsonArray" : true,
      "asJsonObject" : { },
      "asCharacter" : { },
      "jsonObject" : true
    },
    "asLong" : 12345,
    "jsonNull" : true,
    "asJsonPrimitive" : {
      "string" : true,
      "number" : true,
      "boolean" : true,
      "asString" : "...",
      "asDouble" : 12345.0,
      "asFloat" : 12345.0,
      "asJsonNull" : { },
      "jsonPrimitive" : true,
      "asBoolean" : true,
      "asNumber" : { },
      "asJsonArray" : { },
      "asLong" : 12345,
      "jsonNull" : true,
      "asJsonPrimitive" : { },
      "asInt" : 12345,
      "asByte" : 12345,
      "asBigDecimal" : 12345.0,
      "asShort" : 12345,
      "asBigInteger" : 12345,
      "jsonArray" : true,
      "asJsonObject" : { },
      "asCharacter" : { },
      "jsonObject" : true
    },
    "asInt" : 12345,
    "asByte" : 12345,
    "asBigDecimal" : 12345.0,
    "asShort" : 12345,
    "asBigInteger" : 12345,
    "jsonArray" : true,
    "asJsonObject" : { },
    "asCharacter" : { },
    "jsonObject" : true
  },
  "asCharacter" : { },
  "jsonObject" : true
}
            },
            "description" : "Success"
          }
        }
      }
    }
...
    "\/stores/{store}/items/delete/{cartItemId}" : {
      "delete" : {
        "tags" : [ "Stores" ],
        "description" : "",
        "operationId" : "resource_Stores_deleteItemFromCart_DELETE",
        "produces" : [ "application/json" ],
        "parameters" : [
          {
            "name" : "cartItemId",
            "in" : "path",
            "required" : true,
            "type" : "number",
            "description" : ""
          },
          {
            "name" : "store",
            "in" : "path",
            "required" : true,
            "type" : "integer",
            "description" : ""
          }
        ],
        "responses" : {
          "204" : {
            "schema" : {
"description" : "",
"$ref" : "#/definitions/json_JsonObject"
            },
            "headers" : {
            },
            "examples" : {
              "application/json" : {
  "asString" : "...",
  "asDouble" : 12345.0,
  "asFloat" : 12345.0,
  "asJsonNull" : {
    "asString" : "...",
    "asDouble" : 12345.0,
    "asFloat" : 12345.0,
    "asJsonNull" : { },
    "jsonPrimitive" : true,
    "asBoolean" : true,
    "asNumber" : { },
    "asJsonArray" : {
      "asString" : "...",
      "asDouble" : 12345.0,
      "asFloat" : 12345.0,
      "asJsonNull" : { },
      "jsonPrimitive" : true,
      "asBoolean" : true,
      "asNumber" : { },
      "asJsonArray" : { },
      "asLong" : 12345,
      "jsonNull" : true,
      "asJsonPrimitive" : { },
      "asInt" : 12345,
      "asByte" : 12345,
      "asBigDecimal" : 12345.0,
      "asShort" : 12345,
      "asBigInteger" : 12345,
      "jsonArray" : true,
      "asJsonObject" : { },
      "asCharacter" : { },
      "jsonObject" : true
    },
    "asLong" : 12345,
    "jsonNull" : true,
    "asJsonPrimitive" : {
      "string" : true,
      "number" : true,
      "boolean" : true,
      "asString" : "...",
      "asDouble" : 12345.0,
      "asFloat" : 12345.0,
      "asJsonNull" : { },
      "jsonPrimitive" : true,
      "asBoolean" : true,
      "asNumber" : { },
      "asJsonArray" : { },
      "asLong" : 12345,
      "jsonNull" : true,
      "asJsonPrimitive" : { },
      "asInt" : 12345,
      "asByte" : 12345,
      "asBigDecimal" : 12345.0,
      "asShort" : 12345,
      "asBigInteger" : 12345,
      "jsonArray" : true,
      "asJsonObject" : { },
      "asCharacter" : { },
      "jsonObject" : true
    },
    "asInt" : 12345,
    "asByte" : 12345,
    "asBigDecimal" : 12345.0,
    "asShort" : 12345,
    "asBigInteger" : 12345,
    "jsonArray" : true,
    "asJsonObject" : {
      "asString" : "...",
      "asDouble" : 12345.0,
      "asFloat" : 12345.0,
      "asJsonNull" : { },
      "jsonPrimitive" : true,
      "asBoolean" : true,
      "asNumber" : { },
      "asJsonArray" : { },
      "asLong" : 12345,
      "jsonNull" : true,
      "asJsonPrimitive" : { },
      "asInt" : 12345,
      "asByte" : 12345,
      "asBigDecimal" : 12345.0,
      "asShort" : 12345,
      "asBigInteger" : 12345,
      "jsonArray" : true,
      "asJsonObject" : { },
      "asCharacter" : { },
      "jsonObject" : true
    },
    "asCharacter" : { },
    "jsonObject" : true
  },
  "jsonPrimitive" : true,
  "asBoolean" : true,
  "asNumber" : { },
  "asJsonArray" : {
    "asString" : "...",
    "asDouble" : 12345.0,
    "asFloat" : 12345.0,
    "asJsonNull" : {
      "asString" : "...",
      "asDouble" : 12345.0,
      "asFloat" : 12345.0,
      "asJsonNull" : { },
      "jsonPrimitive" : true,
      "asBoolean" : true,
      "asNumber" : { },
      "asJsonArray" : { },
      "asLong" : 12345,
      "jsonNull" : true,
      "asJsonPrimitive" : { },
      "asInt" : 12345,
      "asByte" : 12345,
      "asBigDecimal" : 12345.0,
      "asShort" : 12345,
      "asBigInteger" : 12345,
      "jsonArray" : true,
      "asJsonObject" : { },
      "asCharacter" : { },
      "jsonObject" : true
    },
    "jsonPrimitive" : true,
    "asBoolean" : true,
    "asNumber" : { },
    "asJsonArray" : { },
    "asLong" : 12345,
    "jsonNull" : true,
    "asJsonPrimitive" : {
      "string" : true,
      "number" : true,
      "boolean" : true,
      "asString" : "...",
      "asDouble" : 12345.0,
      "asFloat" : 12345.0,
      "asJsonNull" : { },
      "jsonPrimitive" : true,
      "asBoolean" : true,
      "asNumber" : { },
      "asJsonArray" : { },
      "asLong" : 12345,
      "jsonNull" : true,
      "asJsonPrimitive" : { },
      "asInt" : 12345,
      "asByte" : 12345,
      "asBigDecimal" : 12345.0,
      "asShort" : 12345,
      "asBigInteger" : 12345,
      "jsonArray" : true,
      "asJsonObject" : { },
      "asCharacter" : { },
      "jsonObject" : true
    },
    "asInt" : 12345,
    "asByte" : 12345,
    "asBigDecimal" : 12345.0,
    "asShort" : 12345,
    "asBigInteger" : 12345,
    "jsonArray" : true,
    "asJsonObject" : {
      "asString" : "...",
      "asDouble" : 12345.0,
      "asFloat" : 12345.0,
      "asJsonNull" : { },
      "jsonPrimitive" : true,
      "asBoolean" : true,
      "asNumber" : { },
      "asJsonArray" : { },
      "asLong" : 12345,
      "jsonNull" : true,
      "asJsonPrimitive" : { },
      "asInt" : 12345,
      "asByte" : 12345,
      "asBigDecimal" : 12345.0,
      "asShort" : 12345,
      "asBigInteger" : 12345,
      "jsonArray" : true,
      "asJsonObject" : { },
      "asCharacter" : { },
      "jsonObject" : true
    },
    "asCharacter" : { },
    "jsonObject" : true
  },
  "asLong" : 12345,
  "jsonNull" : true,
  "asJsonPrimitive" : {
    "string" : true,
    "number" : true,
    "boolean" : true,
    "asString" : "...",
    "asDouble" : 12345.0,
    "asFloat" : 12345.0,
    "asJsonNull" : {
      "asString" : "...",
      "asDouble" : 12345.0,
      "asFloat" : 12345.0,
      "asJsonNull" : { },
      "jsonPrimitive" : true,
      "asBoolean" : true,
      "asNumber" : { },
      "asJsonArray" : { },
      "asLong" : 12345,
      "jsonNull" : true,
      "asJsonPrimitive" : { },
      "asInt" : 12345,
      "asByte" : 12345,
      "asBigDecimal" : 12345.0,
      "asShort" : 12345,
      "asBigInteger" : 12345,
      "jsonArray" : true,
      "asJsonObject" : { },
      "asCharacter" : { },
      "jsonObject" : true
    },
    "jsonPrimitive" : true,
    "asBoolean" : true,
    "asNumber" : { },
    "asJsonArray" : {
      "asString" : "...",
      "asDouble" : 12345.0,
      "asFloat" : 12345.0,
      "asJsonNull" : { },
      "jsonPrimitive" : true,
      "asBoolean" : true,
      "asNumber" : { },
      "asJsonArray" : { },
      "asLong" : 12345,
      "jsonNull" : true,
      "asJsonPrimitive" : { },
      "asInt" : 12345,
      "asByte" : 12345,
      "asBigDecimal" : 12345.0,
      "asShort" : 12345,
      "asBigInteger" : 12345,
      "jsonArray" : true,
      "asJsonObject" : { },
      "asCharacter" : { },
      "jsonObject" : true
    },
    "asLong" : 12345,
    "jsonNull" : true,
    "asJsonPrimitive" : { },
    "asInt" : 12345,
    "asByte" : 12345,
    "asBigDecimal" : 12345.0,
    "asShort" : 12345,
    "asBigInteger" : 12345,
    "jsonArray" : true,
    "asJsonObject" : {
      "asString" : "...",
      "asDouble" : 12345.0,
      "asFloat" : 12345.0,
      "asJsonNull" : { },
      "jsonPrimitive" : true,
      "asBoolean" : true,
      "asNumber" : { },
      "asJsonArray" : { },
      "asLong" : 12345,
      "jsonNull" : true,
      "asJsonPrimitive" : { },
      "asInt" : 12345,
      "asByte" : 12345,
      "asBigDecimal" : 12345.0,
      "asShort" : 12345,
      "asBigInteger" : 12345,
      "jsonArray" : true,
      "asJsonObject" : { },
      "asCharacter" : { },
      "jsonObject" : true
    },
    "asCharacter" : { },
    "jsonObject" : true
  },
  "asInt" : 12345,
  "asByte" : 12345,
  "asBigDecimal" : 12345.0,
  "asShort" : 12345,
  "asBigInteger" : 12345,
  "jsonArray" : true,
  "asJsonObject" : {
    "asString" : "...",
    "asDouble" : 12345.0,
    "asFloat" : 12345.0,
    "asJsonNull" : {
      "asString" : "...",
      "asDouble" : 12345.0,
      "asFloat" : 12345.0,
      "asJsonNull" : { },
      "jsonPrimitive" : true,
      "asBoolean" : true,
      "asNumber" : { },
      "asJsonArray" : { },
      "asLong" : 12345,
      "jsonNull" : true,
      "asJsonPrimitive" : { },
      "asInt" : 12345,
      "asByte" : 12345,
      "asBigDecimal" : 12345.0,
      "asShort" : 12345,
      "asBigInteger" : 12345,
      "jsonArray" : true,
      "asJsonObject" : { },
      "asCharacter" : { },
      "jsonObject" : true
    },
    "jsonPrimitive" : true,
    "asBoolean" : true,
    "asNumber" : { },
    "asJsonArray" : {
      "asString" : "...",
      "asDouble" : 12345.0,
      "asFloat" : 12345.0,
      "asJsonNull" : { },
      "jsonPrimitive" : true,
      "asBoolean" : true,
      "asNumber" : { },
      "asJsonArray" : { },
      "asLong" : 12345,
      "jsonNull" : true,
      "asJsonPrimitive" : { },
      "asInt" : 12345,
      "asByte" : 12345,
      "asBigDecimal" : 12345.0,
      "asShort" : 12345,
      "asBigInteger" : 12345,
      "jsonArray" : true,
      "asJsonObject" : { },
      "asCharacter" : { },
      "jsonObject" : true
    },
    "asLong" : 12345,
    "jsonNull" : true,
    "asJsonPrimitive" : {
      "string" : true,
      "number" : true,
      "boolean" : true,
      "asString" : "...",
      "asDouble" : 12345.0,
      "asFloat" : 12345.0,
      "asJsonNull" : { },
      "jsonPrimitive" : true,
      "asBoolean" : true,
      "asNumber" : { },
      "asJsonArray" : { },
      "asLong" : 12345,
      "jsonNull" : true,
      "asJsonPrimitive" : { },
      "asInt" : 12345,
      "asByte" : 12345,
      "asBigDecimal" : 12345.0,
      "asShort" : 12345,
      "asBigInteger" : 12345,
      "jsonArray" : true,
      "asJsonObject" : { },
      "asCharacter" : { },
      "jsonObject" : true
    },
    "asInt" : 12345,
    "asByte" : 12345,
    "asBigDecimal" : 12345.0,
    "asShort" : 12345,
    "asBigInteger" : 12345,
    "jsonArray" : true,
    "asJsonObject" : { },
    "asCharacter" : { },
    "jsonObject" : true
  },
  "asCharacter" : { },
  "jsonObject" : true
}
            },
            "description" : "Success"
          }
        }
      }
    }
  }
}